+++ tests/test_diff_text.py
@@ -1196,6 +1196,10 @@ class TestDeepDiffText:
         result = {'values_changed': {'root[4]': {'new_value': 'now', 'old_value': now}}}
         assert result == ddiff
 
+    def test_ignore_type_in_groups_float_vs_decimal(self):
+        diff = DeepDiff(float('0.1'), Decimal('0.1'), ignore_type_in_groups=[(float, Decimal)], significant_digits=2)
+        assert not diff
+
     @pytest.mark.parametrize("t1, t2, significant_digits, result", [
         ([0.1], [Decimal('0.10')], 55,
             {'values_changed': {'root[0]': {'new_value': Decimal('0.10'), 'old_value': 0.1}}}),  # Due to floating point arithmetics with high significant digits.
@@ -1395,6 +1399,30 @@ class TestDeepDiffText:
         ddiff = DeepDiff(t1, t2, exclude_regex_paths=[r"root\[\d+\]\['b'\]"])
         result = {}
         assert result == ddiff
+
+    def test_include_obj_callback(self):
+        def include_obj_callback(obj, path):
+            return True if "include" in path or isinstance(obj, int) else False
+
+        t1 = {"x": 10, "y": "b", "z": "c", "include_me": "a"}
+        t2 = {"x": 10, "y": "c", "z": "b", "include_me": "b"}
+        ddiff = DeepDiff(t1, t2, include_obj_callback=include_obj_callback)
+        result = {'values_changed': {"root['include_me']": {'new_value': "b", 'old_value': "a"}}}
+        assert result == ddiff
+        assert {"root['include_me']"} == ddiff.affected_paths
+        assert {"include_me"} == ddiff.affected_root_keys
+
+    def test_include_obj_callback_strict(self):
+        def include_obj_callback_strict(obj, path):
+            return True if isinstance(obj, int) and obj > 10 else False
+
+        t1 = {"x": 11, "y": 10, "z": "c"}
+        t2 = {"x": 12, "y": 12, "z": "c"}
+        ddiff = DeepDiff(t1, t2, include_obj_callback_strict=include_obj_callback_strict)
+        result = {'values_changed': {"root['x']": {'new_value': 12, 'old_value': 11}}}
+        assert result == ddiff
+        assert {"root['x']"} == ddiff.affected_paths
+        assert {"x"} == ddiff.affected_root_keys
 
     def test_skip_exclude_obj_callback(self):
         def exclude_obj_callback(obj, path):